# lib/api-zod/src/generated/types/updateBlogPostRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBlogPostRequestStatus } from "./updateBlogPostRequestStatus";

export interface UpdateBlogPostRequest {
  /**
   * Blog post title
   * @minLength 5
   * @maxLength 255
   */
  title?: string;
  /**
   * URL-friendly slug
   * @minLength 1
   * @maxLength 255
   * @pattern ^[a-z0-9-]+$
   */
  slug?: string;
  /**
   * Blog post content (rich text)
   * @minLength 50
   */
  content?: string;
  /**
   * SEO meta description (optional)
   * @maxLength 160
   */
  metaDescription?: string;
  /** Whether this is a featured post */
  isFeatured?: boolean;
  /** Publication status */
  status?: UpdateBlogPostRequestStatus;
}